streamlit_docker/requirements.txt
streamlit==1.31.1
pymongo
opencv-python==4.9.0.80
matplotlib==3.8.2
pandas==2.2.0
numpy>=1.19.3
Pillow>=7.1.0
PyYAML>=5.1
